Changing url validation const and using it in tests without returning side effects
