+++ src/components/Collaborative/hooks/useFloatingLinkState.tsx
@@ -18,9 +18,7 @@ import { useTranslation } from "next-i18next";
 import { VisualEditorContext } from "../VisualEditorProvider";
 import useLinkShortcut from "./useLinkShortcut";
 import { uniqueId } from "remirror";
-
-export const URL_PATTERN =
-    /^(ftp|http|https):\/\/[^ "]+\.[a-zA-Z]{2,}(\/|\?|#|$)/;
+import { validateFloatingLink } from "../../../utils/ValidateFloatingLink";
 
 function useFloatingLinkState() {
     const { t } = useTranslation();
@@ -53,13 +51,6 @@ function useFloatingLinkState() {
     }, [isEditing, setIsEditing, updateReason.doc, updateReason.selection]);
 
     useEffect(() => setHref(url), [url]);
-
-    const validateFloatingLink = useCallback(() => {
-        // TODO: use a library or service that maintains a comprehensive list of valid TLDs
-        if (!URL_PATTERN.test(href)) {
-            throw new Error(t("sourceForm:errorMessageValidURL"));
-        }
-    }, [href, t]);
 
     const updateFloatingLink = useCallback(
         (id) => {
