+++ src/components/Form/FormField.ts
@@ -3,7 +3,7 @@ import { RegisterOptions } from "react-hook-form";
 import { EditorParser } from "../../../lib/editor-parser";
 import { ReviewTaskMachineContextReviewData } from "../../../server/review-task/dto/create-review-task.dto";
 import { Roles } from "../../types/enums";
-import { URL_PATTERN } from "../Collaborative/hooks/useFloatingLinkState";
+import { URL_PATTERN } from "../../utils/ValidateFloatingLink";
 
 export type FormField = {
     fieldName: string;
